+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Rollout workflow may optionally include phasing of heat buffers for handling peak day constraints and demands
 - Inclusion of maximum allowed yearly pipe placement length constraint for the rollout workflow
 - Adding constraint so that all heating demands are placed at the end of the simulation
-
+- Gas Boiler with 2 ports (no gas port) is supported for DTK and its cost attributes are supported in Mesido
 
 ## Changed
 - Minimize TCO objective in the grow_workflow is now only based on capex and opex that can be influenced.